requirements.txt
Flask
twilio
gunicorn
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
dateparser
requests
SpeechRecognition
gTTS